+++ server.js
@@ -2,13 +2,15 @@ import { config } from "./config/index.js";
 import app from "./app.js";
 import { appLoader } from "./src/core/AppLoader.js";
 import { initEmailTemplates } from "./scripts/init_email_templates.js";
+import { migratePreferences } from "./scripts/migrate_preferences.js";
 // Start Email Worker
 import "./services/emailWorker.js";
 
 // Initialize Services
 Promise.all([
   appLoader.init(),
-  initEmailTemplates()
+  initEmailTemplates(),
+  migratePreferences()
 ]).then(() => {
   console.log("Core services initialized");